requirements.txt
streamlit
cohere
pinecone-client
PyMuPDF
pickle-mixin
python-dotenv
utils